chore: initial server logic separation
